+++ doc/Tree/pruning.jl
@@ -1,4 +1,4 @@
-using .MCPhylo
+using MCPhylo
 using Test
 
 tree = MCPhylo.parsing_newick_string("(A,B,(C,(D,E)F)G)H;")