+++ src/Sampler/PNUTS.jl
@@ -20,7 +20,6 @@ mutable struct PNUTSTune <: SamplerTune
   moves::Int
 
 
-
   PNUTSTune() = new()
 
   function PNUTSTune(x::Vector{T}, epsilon::Float64, logfgrad::Union{Function, Missing};
@@ -69,31 +68,27 @@ function mlogpdfgrad!(block::SamplingBlock, x::T, sz::Int64, ll::Bool=false, gr:
   end
   lp, grad
 end
-
 sample!(v::PNUTSVariate; args...) = sample!(v, v.tune.logfgrad; args...)
 
-function sample!(v::PNUTSVariate, logfgrad::Function; adapt::Bool=false)::PNUTSVariate
+function sample!(v::PNUTSVariate, logfgrad::Function; adapt::Bool=false)
 
   tune = v.tune
   setadapt!(v, adapt)
-
   if tune.adapt
     tune.m += 1
-    nuts_sub!(v, logfgrad)
+    nuts_sub!(v, tune.epsilon, logfgrad)
     p = 1.0 / (tune.m + tune.t0)
     tune.Hbar = (1.0 - p) * tune.Hbar +
                 p * (tune.target - tune.alpha / tune.nalpha)
     tune.epsilon = exp(tune.mu - sqrt(tune.m) * tune.Hbar / tune.gamma)
-
     p = tune.m^-tune.kappa
     tune.epsilonbar = exp(p * log(tune.epsilon) +
                           (1.0 - p) * log(tune.epsilonbar))
   else
     if (tune.m > 0) tune.epsilon = tune.epsilonbar end
 
-    nuts_sub!(v, logfgrad)
+    nuts_sub!(v, tune.epsilon, logfgrad)
   end
-
   v
 end
 
@@ -110,26 +105,25 @@ end
 
 
 
-function nuts_sub!(v::PNUTSVariate, logfgrad::Function)::PNUTSVariate
-  mt::Node = deepcopy(v[1])
-  epsilon::Float64 = v.tune.epsilon
-  delta::Float64 = v.tune.delta
+function nuts_sub!(v::PNUTSVariate, epsilon::Float64, logfgrad::Function)
 
-  nl::Int64 = size(mt)[1]-1
+  mt = v.value[1]
+  nl = size(mt)[1]-1
+  delta = v.tune.delta
+
   r = randn(nl)
   g = zeros(nl)
 
-  nni = zero(Int64)
+  x, r , logf, grad, nni = refraction(mt, r, 1, g, epsilon, logfgrad, delta, nl)
 
-  x::Node, r::Vector{Float64}, logf::Float64, grad::Vector{Float64}, nni = refraction(mt, r, 1, g, epsilon, logfgrad, delta, nl)
+  lu = log(rand())
 
+  logp0 = logf - 0.5 * dot(r)
+  logu0 = logp0 + lu
+  rminus = rplus = r
+  gradminus = gradplus = grad
 
-
-  logp0::Float64 = logf - 0.5 * dot(r)
-  logu0::Float64 = logp0 + log(rand())
-  rminus::Vector{Float64} = rplus::Vector{Float64} = r
-  gradminus::Vector{Float64} = gradplus::Vector{Float64} = grad
-  xminus::Node = xplus::Node = x
+  xminus = xplus = x
 
   j = 0
   n = 1
@@ -142,19 +136,19 @@ function nuts_sub!(v::PNUTSVariate, logfgrad::Function)::PNUTSVariate
     if pm == -1
 
       xminus, rminus, gradminus, _, _, _, xprime, nprime, sprime, alpha,
-        nalpha, nni1 = buildtree(xminus, rminus, gradminus, pm, j, epsilon, logfgrad,
-                           logp0, logu0, delta, nl)
+        nalpha, nni1, lpp = buildtree(xminus, rminus, gradminus, pm, j, epsilon, logfgrad,
+                           logp0, logu0, delta, nl,lu)
 
     else
 
-      _, _, _, xplus, rplus, gradplus, xprime, nprime, sprime, alpha, nalpha, nni1 =
+      _, _, _, xplus, rplus, gradplus, xprime, nprime, sprime, alpha, nalpha, nni1, lpp =
         buildtree(xplus, rplus, gradplus, pm, j, epsilon, logfgrad, logp0,
-                  logu0, delta, nl)
+                  logu0, delta, nl, lu)
 
     end#if pm
 
     if sprime && rand() < nprime / n
-        v.value[1] = xprime
+        v.value[1]= xprime
     end
     j += 1
     n += nprime
@@ -164,26 +158,25 @@ function nuts_sub!(v::PNUTSVariate, logfgrad::Function)::PNUTSVariate
   end
   v.tune.moves += nni
   v
-
 end
 
 
 function refraction(v::T, r::Vector{Float64}, pm::Int64,
                     grad::Vector{Float64}, epsilon::Float64, logfgrad::Function,
-                    delta::Float64, sz::Int64)::Tuple{T, Vector{Float64}, Float64, Vector{Float64}, Int64} where T<:Node
+                    delta::Float64, sz::Int64)  where T<:Node
 
-    v1::T = deepcopy(v)
+    v1 = deepcopy(v)
 
-    ref_r::Vector{Float64} = pm*r
+    ref_r = pm*r
 
-    blenvec::Vector{Float64} = get_branchlength_vector(v1)
-    fac::Vector{Float64} = scale_fac.(blenvec, delta)
+    blenvec = get_branchlength_vector(v1)
+    fac = scale_fac.(blenvec, delta)
+
 
     ref_r = @. ref_r + (epsilon * 0.5) * (grad * fac)
-
     tmpB = @. blenvec + (epsilon * ref_r)
 
-    nni = zero(Int64)
+    nni = 0
 
     if minimum(tmpB) <= 0
         v1, tmpB, ref_r, nni = ref_NNI(v1, tmpB, ref_r, epsilon, blenvec, delta, logfgrad, sz)
@@ -195,8 +188,8 @@ function refraction(v::T, r::Vector{Float64}, pm::Int64,
     set_branchlength_vector!(v1, blenvec)
 
 
+    logf, grad = logfgrad(v1, sz, true, true)
 
-    logf::Float64, grad = logfgrad(v1, sz, true, true)
     fac = scale_fac.(blenvec, delta)
     ref_r = @. ref_r + (epsilon * 0.5) * (grad * fac)
 
@@ -206,16 +199,12 @@ function refraction(v::T, r::Vector{Float64}, pm::Int64,
 end
 
 
-
-
-
 function ref_NNI(v::T, tmpB::Vector{Float64}, r::Vector{Float64}, epsilon::Float64, blv::Vector{Float64},
-                 delta::Float64, logfgrad::Function, sz::Int64)::Tuple{T,Vector{Float64},Vector{Float64},Int64}  where T<:Node
+                 delta::Float64, logfgrad::Function, sz::Int64)  where T<:Node
 
-  intext::Vector{Int64} = internal_external(v)
-  t = zero(Float64)
-  nni = zero(Int64)
-
+  intext = internal_external(v)
+  t = 0.0
+  nni = 0
 
   while minimum(tmpB)<=0.0
 
@@ -224,26 +213,27 @@ function ref_NNI(v::T, tmpB::Vector{Float64}, r::Vector{Float64}, epsilon::Float
 
      temp = epsilon-t+timelist[ref_index]
      blv = @. blv + (temp * r)
+
      r[ref_index] *= -1.0
 
      if intext[ref_index] == 1
 
-       blv1::Vector{Float64} = molifier.(blv, delta)
+       blv1 =molifier.(blv, delta)
        set_branchlength_vector!(v, blv1)
 
        # use thread parallelism
        res_before = @spawn logfgrad(v, sz, true, false) # still with molified branch length
 
-       v_copy::T = deepcopy(v)
-       tmp_NNI_made::Int64 = NNI!(v_copy, ref_index)
+       v_copy = deepcopy(v)
+       tmp_NNI_made = NNI!(v_copy, ref_index)
 
        # fetch the results from the parallel part
-       U_before_nni, _ = fetch(res_before)::Tuple{Float64, Vector{Float64}}
+       U_before_nni, _ = fetch(res_before)
        U_before_nni *= -1
 
        if tmp_NNI_made != 0
 
-            U_after_nni, _ = logfgrad(v_copy, sz, true, false)::Tuple{Float64, Vector{Float64}}
+            U_after_nni, _ = logfgrad(v_copy, sz, true, false)
             U_after_nni *= -1
             delta_U::Float64 = 2.0*(U_after_nni - U_before_nni)
             my_v::Float64 = r[ref_index]^2
@@ -268,42 +258,42 @@ end
 function buildtree(x::T, r::Vector{Float64},
                    grad::Vector{Float64}, pm::Int64, j::Integer,
                    epsilon::Float64, logfgrad::Function, logp0::Real, logu0::Real,
-                   delta::Float64, sz::Int64)::Tuple{T, Vector{Float64}, Vector{Float64}, T, Vector{Float64},Vector{Float64},T,Int64,Bool,Float64,Int64,Int64} where T<:Node
+                   delta::Float64, sz::Int64, lu::Float64)  where T<:Node
 
 
   if j == 0
 
-    xprime::T, rprime::Vector{Float64}, logfprime::Float64, gradprime::Vector{Float64}, nni::Int64 = refraction(x, r, pm, grad, epsilon,
+    xprime, rprime, logfprime, gradprime, nni = refraction(x, r, pm, grad, epsilon,
                                           logfgrad, delta, sz)
 
-    logpprime::Float64 = logfprime - 0.5 * dot(rprime)
+    logpprime = logfprime - 0.5 * dot(rprime)
 
-    nprime::Int64 = Int(logu0 < logpprime)
+    nprime = Int(logu0 < logpprime)
 
-    sprime::Bool = logu0 < logpprime + 1000.0
-    xminus::T = xplus::T = xprime
-    rminus::Vector{Float64} = rplus::Vector{Float64} = rprime
-    gradminus::Vector{Float64} = gradplus::Vector{Float64} = gradprime
-    alphaprime::Float64 = min(1.0, exp(logpprime - logp0))
+    sprime = logu0 < logpprime + 1000.0
+    xminus = xplus = xprime
+    rminus = rplus = rprime
+    gradminus = gradplus = gradprime
+    alphaprime = min(1.0, exp(logpprime - logp0))
     nalphaprime = 1
 
   else
     xminus, rminus, gradminus, xplus, rplus, gradplus, xprime, nprime, sprime,
-      alphaprime, nalphaprime, nni = buildtree(x, r, grad, pm, j - 1, epsilon,
-                                          logfgrad, logp0, logu0, delta,  sz)
+      alphaprime, nalphaprime, nni, logpprime = buildtree(x, r, grad, pm, j - 1, epsilon,
+                                          logfgrad, logp0, logu0, delta,  sz, lu)
     if sprime
 
       if pm == -1
 
         xminus, rminus, gradminus, _, _, _, xprime2, nprime2, sprime2,
-          alphaprime2, nalphaprime2 , nni = buildtree(xminus, rminus, gradminus, pm,
+          alphaprime2, nalphaprime2 , nni, logpprime= buildtree(xminus, rminus, gradminus, pm,
                                                 j - 1, epsilon, logfgrad, logp0,
-                                                logu0, delta,  sz)
+                                                logu0, delta,  sz, lu)
       else
         _, _, _, xplus, rplus, gradplus, xprime2, nprime2, sprime2,
-          alphaprime2, nalphaprime2, nni = buildtree(xplus, rplus, gradplus, pm,
+          alphaprime2, nalphaprime2, nni, logpprime = buildtree(xplus, rplus, gradplus, pm,
                                                 j - 1, epsilon, logfgrad, logp0,
-                                                logu0, delta, sz)
+                                                logu0, delta, sz,lu)
       end # if pm
 
       if rand() < nprime2 / (nprime + nprime2)
@@ -317,14 +307,13 @@ function buildtree(x::T, r::Vector{Float64},
   end #if j
 
   xminus, rminus, gradminus, xplus, rplus, gradplus, xprime, nprime, sprime,
-    alphaprime, nalphaprime, nni
+    alphaprime, nalphaprime, nni, logpprime
 end
 
 
 function nouturn(xminus::T, xplus::T,
-                rminus::Vector{Float64}, rplus::Vector{Float64}, gradminus::Vector{Float64}, gradplus::Vector{Float64},
-                epsilon::Float64, logfgrad::Function, delta::Float64, sz::Int64, j::Int64)::Bool  where T<:Node
-
+                rminus::Vector{Float64}, rplus::Vector{Float64}, gradminus::Vector{Float64},gradplus::Vector{Float64},
+                epsilon::Float64, logfgrad::Function, delta::Float64, sz::Int64, j::Int64)  where T<:Node
 
         curr_l, curr_h = BHV_bounds(xminus, xplus)
 
@@ -344,43 +333,36 @@ end
 
 #################### Auxilliary Functions ####################
 
-function nutsepsilon(x::T, logfgrad::Function, delta::Float64)::Float64  where T<:Node
+function nutsepsilon(x::Node, logfgrad::Function, delta::Float64)
 
-  x0::T = deepcopy(x)
+  x0 = deepcopy(x)
   n = size(x)[1] - 1
 
   _, r0, logf0, grad0,_ = refraction(x0, randn(n), 1, zeros(n), 0.0, logfgrad, delta, n)
 
   x0 = deepcopy(x)
   epsilon = 1.0
-
   _, rprime, logfprime, gradprime,_ = refraction(x0, r0, 1, grad0,  epsilon, logfgrad, delta, n)
 
   prob = exp(logfprime - logf0 - 0.5 * (dot(rprime) - dot(r0)))
 
   pm = 2 * (prob > 0.5) - 1
-
   while prob^pm > 0.5^pm
-   epsilon *= 2.0^pm
-   x0 = deepcopy(x)
-   _, rprime, logfprime, _ ,_ = refraction(x0, r0, 1, grad0, epsilon, logfgrad, delta, n)
-   prob = exp(logfprime - logf0 - 0.5 * (dot(rprime) - dot(r0)))
-  end
+    epsilon *= 2.0^pm
+    x0 = deepcopy(x)
+    _, rprime, logfprime, _ ,_ = refraction(x0, r0, 1, grad0, epsilon, logfgrad, delta, n)
 
-  println("eps ",epsilon)
+    prob = exp(logfprime - logf0 - 0.5 * (dot(rprime) - dot(r0)))
+
+  end
 
   epsilon
 end
 
-@inline function scale_fac(x::T, delta::T)::T where T<:Real
+@inline function scale_fac(x::T, delta::T) where T<:Float64
   x < delta ? x/delta : 1.0
 end
 
-"""
-    molifier(x::Float64, delta::Float64)::Float64
-
-documentation
-"""
-@inline function molifier(x::T, delta::T)::T where T <: Real
+@inline function molifier(x::Float64, delta::Float64)::Float64
     x >= delta ? x : (x^2+delta^2)/(2.0*delta)
 end # function