+++ src/model/simulation.jl
@@ -40,7 +40,12 @@ end
 function gradlogpdf!(m::Model, x::AbstractVector{T}, block::Integer=0,
                       transform::Bool=false; dtype::Symbol=:forward) where {T<:Real}
   f = x -> logpdf!(m, x, block, transform)
-  gradient(f, convert(Vector{T}, x), dtype)
+  if dtype == :Zygote
+    Zygote.gradient(f, x)
+  else
+    FiniteDiff.finite_difference_gradient(f, x)
+  end
+
 end
 
 
@@ -69,8 +74,8 @@ function logpdf(m::Model, x::AbstractArray{T}, block::Integer=0,
 end
 
 
-function logpdf!(m::Model, x::Node, block::Integer=0,
-                  transform::Bool=false)
+function logpdf!(m::Model, x::N, block::Integer=0,
+                  transform::Bool=false) where N<:GeneralNode
   params = keys(m, :block, block)
   targets = keys(m, :target, block)
   m[params] = relist(m, x, params, transform)
@@ -84,32 +89,41 @@ function logpdf!(m::Model, x::Node, block::Integer=0,
   lp
 end
 
-function gradlogpdf!(m::Model, x::AbstractArray{T}, block::Integer=0,transform::Bool=false)where {T<:AbstractNode}
+function gradlogpdf!(m::Model, x::AbstractArray{T}, block::Integer=0,transform::Bool=false) where T<:GeneralNode
   gradlogpdf!(m, x, block, transform)
 end
 
 
-function gradlogpdf!(m::Model, x::Node, block::Integer=0,transform::Bool=false)::Tuple{Float64, Vector{Float64}}
+function gradlogpdf(m::Model, targets::Array{Symbol, 1})::Tuple{Float64, Array{Float64}}
+  vp = 0.0
+  gradp = Array[]
+  for key in targets
+    node = m[key]
+    update!(node, m)
+    v, grad = gradlogpdf(node)
+    vp += v
+    push!(gradp, grad)
+  end
+  vp, .+(gradp...)
+end
+
+function gradlogpdf!(m::Model, x::N, block::Integer=0,transform::Bool=false)::Tuple{Float64, Vector{Float64}} where N<:GeneralNode
   params = keys(m, :block, block)
   targets = keys(m, :target, block)
   m[params] = relist(m, x, params, transform)
 
 
-
   # use thread parallelism
-  # prior
-  prior_res = @spawn gradlogpdf(m[params[1]], x)
-  #prior_res =  gradlogpdf(m[params[1]], x)
-
   # likelihood
-  v, grad = gradlogpdf(m[targets[1]])
+  lik_res = @spawn gradlogpdf(m, targets)
+
+  # prior
+  vp, gradp =  gradlogpdf(m[params[1]], x)
 
   # get results from threads
-  vp, gradp = fetch(prior_res)
+  v, grad = fetch(lik_res)
 
-
-
-  v+vp, grad.+gradp
+  vp+v, gradp.+grad
 end
 
 function logpdf!(m::Model, x::AbstractArray{T}, block::Integer=0,
@@ -167,10 +181,11 @@ end
 function unlist(m::Model, monitoronly::Bool)
   f = function(key)
     node = m[key]
-    lvalue = unlist(node)
+    lvalue = isa(node, AbstractTreeStochastic) ? unlist_tree(node) : unlist(node)
     monitoronly ? lvalue[node.monitor] : lvalue
   end
-  vcat(map(f, keys(m, :dependent))..., m.likelihood)
+  r = vcat(map(f, keys(m, :dependent))..., m.likelihood)
+  r
 end
 
 function unlist(m::Model, nodekeys::Vector{Symbol}, transform::Bool=false)
@@ -184,9 +199,9 @@ function relist(m::Model, x::AbstractArray{T}, block::Integer=0,
 end
 
 function relist(m::Model, x::AbstractArray{T}, block::Integer=0,
-                transform::Bool=false) where {T<:AbstractNode}
+               transform::Bool=false) where {T<:GeneralNode}
 
-  relist(m, x, keys(m, :block, block), transform)
+ relist(m, x, keys(m, :block, block), transform)
 end
 
 
@@ -197,7 +212,9 @@ function relist(m::Model, x::AbstractArray{T},
   N = length(x)
   offset = 0
   for key in nodekeys
+
     value, n = relistlength(m[key], view(x, (offset + 1):N), transform)
+
     values[key] = value
     offset += n
   end
@@ -208,10 +225,8 @@ end
 
 
 
-function relist(m::Model, x::Node,
-                nodekeys::Vector{Symbol}, transform::Bool=false)
+function relist(m::Model, x::N, nodekeys::Vector{Symbol}, transform::Bool=false) where N<:GeneralNode
   values = Dict{Symbol,Any}()
-  #N = length(x)
   offset = 0
   for key in nodekeys
     value, n = relistlength(m[key], x, transform)
@@ -239,7 +254,7 @@ function assign!(m::Model, key::Symbol, value::T) where T <: Real
   m[key].value = value
 end
 
-function assign!(m::Model, key::Symbol, value::T) where T <:AbstractNode
+function assign!(m::Model, key::Symbol, value::T) where T <:GeneralNode
   m[key].value = value
 end